Refactored FirmwareUpgradeManager's main Public API functions into their own file
As well as helping improve the overall size of the main file, it's helpful for users of this library to figure out what are the main calls they should be interacting with.
